Split frontend debt ratchet batch 2
