+++ __tests__/scripts/debt-ratchet.test.ts
@@ -306,13 +306,9 @@ describe("debt-ratchet check mode", () => {
     expect(check.stdout).toMatch(
       /^oversized_files\s+baseline\s+2\s+actual\s+2\s+ok$/m
     );
-    expect(check.stdout).toMatch(/^\s+breakdown:$/m);
-    expect(check.stdout).toMatch(
-      /^\s+app_source\s+baseline\s+1\s+actual\s+1$/m
-    );
-    expect(check.stdout).toMatch(
-      /^\s+wp_migrated\s+baseline\s+1\s+actual\s+1$/m
-    );
+    expect(check.stdout).not.toContain("breakdown:");
+    expect(check.stdout).not.toContain("app_source");
+    expect(check.stdout).not.toContain("wp_migrated");
 
     const unfiltered = runRatchet(root, ["--details", "oversized_files"]);
     expect(unfiltered.status).toBe(0);
