+++ components/eula/EULACommunicationsSection.tsx
@@ -1,0 +1,29 @@
+export default function EULACommunicationsSection() {
+  return (
+    <li>
+      <p>
+        <strong>Our Platform, Our House Rules III</strong>
+      </p>
+      <p>
+        <a
+          href="https://twitter.com/punk6529"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          @punk6529
+        </a>{" "}
+        tweets a lot and has a lot of ideas that may change from time to time.
+      </p>
+      <p>
+        You should have no expectation whatsoever that what he (or other team
+        members) are going to tweet or otherwise communicate is going to match
+        your current or future understanding of the 6529 NFTs.
+      </p>
+      <p>
+        His, or any other person or team member's, retweets, likes, twitter or
+        discord communications, podcasts, essays or other communications are not
+        endorsements nor do they reflect changes in Terms.
+      </p>
+    </li>
+  );
+}
