+++ components/eula/EULAIntroSections.tsx
@@ -1,0 +1,113 @@
+export default function EULAIntroSections() {
+  return (
+    <>
+      <li>
+        <p>
+          <strong>Abbreviations Make Documents More Readable</strong>
+        </p>
+        <p>These are the abbreviations used in this document.</p>
+        <p>
+          "Terms" or "EULA": End User License Agreement (EULA)
+          <br />
+          "We": 6529 Collection LLC
+          <br />
+          "6529 NFTs": The Memes, 6529 Gradient, Meme Lab, GenMemes, 6529 Intern
+          and other NFTs that may be added from time to time.
+          <br />
+          "You": An adult, at least 18 years of age, who is not subject to
+          sanctions by the US government.
+          <br />
+          "Third Parties": Everyone else who is not "We" or "You"
+          <br />
+          "Our App": The 6529Mobile application and any associated services we
+          may offer
+          <br />
+          "Our Platform": Our App, the website located at{" "}
+          <a href="https://6529.io" target="_blank" rel="noopener noreferrer">
+            6529.io
+          </a>
+          , any websites hosted at sub-domains of{" "}
+          <a href="https://6529.io" target="_blank" rel="noopener noreferrer">
+            6529.io
+          </a>
+          , including hosted minting or primary sales pages, any primary mints
+          or sales directly from our smart contracts, any mobile or metaverse
+          applications we may make, and any content (data, descriptions or
+          otherwise) on our website or decentralized file storage platforms like
+          IPFS or Arweave.
+          <br />
+          "Not Our Platform": Everything else that is not included in "Our
+          Platform" including your Ethereum wallet, NFT marketplaces, and
+          publicly accessible secondary functions on our smart contracts, such
+          as token transfers.
+        </p>
+      </li>
+      <li>
+        <p>
+          <strong>The Underlying Principles Of These Terms</strong>
+        </p>
+        <p>There are two underlying principles of these terms:</p>
+        <p>
+          Our main ongoing activity is primary minting CC0 NFTs (The Memes, Meme
+          Lab, GenMemes are CC0) with a mission-based focus at fairly affordable
+          prices.
+        </p>
+        <p>
+          Once the NFTs have been primary minted:
+          <br />
+          a) You and Third Parties can do whatever you want with the NFT
+          <br />
+          and
+          <br />
+          b) We do not have any further obligations toward you
+        </p>
+        <p>
+          Most of the Clauses in the Terms relate to making (b) absolutely
+          clear.
+        </p>
+        <p>
+          We do not have control of what someone does with The Memes, Meme Lab
+          and GenMemes NFTs after minting, either at the level of the token or
+          the art.
+        </p>
+        <p>
+          "Our Platforms" provide a (hopefully) convenient way to see
+          information about 6529 NFTs and their collectors, but they are not the
+          token nor the art. We do not make any claims on the token post-minting
+          or on the art for the CC0 6529 NFTs, nor do we take any responsibility
+          for either.
+        </p>
+      </li>
+      <li>
+        <p>
+          <strong>Our Platform, Our House Rules I</strong>
+        </p>
+        <p>
+          Agreeing to these terms is the only way you are allowed to use
+          6529Mobile. If you do not agree with our Terms, that is fine, but you
+          cannot use 6529Mobile.
+        </p>
+      </li>
+      <li>
+        <p>
+          <strong>Our Platform, Our House Rules II</strong>
+        </p>
+        <p>
+          We can terminate or pause your access to Our Platform at any time, for
+          any reason.
+        </p>
+        <p>
+          We strive for a positive community, and that means no abusive behavior
+          or posting objectionable user-generated content. Please report any
+          concerns to us at <a href="mailto:support@6529.io">support@6529.io</a>
+          . Our policy is to respond to these reports within 24 hours, which
+          includes removing content or blocking users as warranted.
+        </p>
+        <p>
+          Please note: We monitor material posted to the app to filter for
+          objectionable material.
+        </p>
+      </li>
+    </>
+  );
+}
