+++ scripts/debt-ratchet-baseline.json
@@ -4,28 +4,20 @@
   "counts": {
     "any_casts": 0,
     "todo_comments": 0,
-    "oversized_files": 16,
+    "oversized_files": 8,
     "legacy_wordpress_runtime": 0,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
   },
   "oversized_file_allowlist": [
-    "components/about/AboutTermsOfService.tsx",
-    "components/auth/SeizeConnectProvider.tsx",
-    "components/distribution-plan-tool/build-phases/build-phase/form/BuildPhaseFormConfigModal.tsx",
-    "components/distribution-plan-tool/review-distribution-plan/table/ReviewDistributionPlanTableSubscriptionFooter.tsx",
     "components/drops/view/item/content/attachments/DropAttachmentDisplay.tsx",
     "components/drops/view/item/content/media/SeizeVideoPlayer.tsx",
-    "components/eula/EULAModal.tsx",
     "components/manifold-minting/ManifoldMinting.tsx",
     "components/meme-calendar/meme-calendar.helpers.tsx",
     "components/memelab/MemeLabPage.tsx",
     "components/nextGen/admin/NextGenAdmin.tsx",
     "components/nft-transfer/TransferModal.tsx",
-    "components/react-query-wrapper/ReactQueryWrapper.tsx",
-    "components/rememes/RememePage.tsx",
-    "components/user/collected/UserPageCollected.tsx",
-    "helpers/Helpers.ts"
+    "components/rememes/RememePage.tsx"
   ]
 }
